refactor(server): peel dashboard analytics routes into routers/analytics.py (router-split, round-5 #51 / R5-25 PR19)
Eighth router peel. Moves the 5 read-only dashboard aggregate endpoints —
/api/stats (+ real disk usage, top tags, registered project name), /api/tags,
/api/collections, /api/duration-by-lang, /api/size-by-ext — into a leaf
routers/analytics.py, along with their two group-local helpers
_current_project_registry_name and _thumb_url. config.BASE_DIR replaces
server.ROOT for the get_stats disk-usage fallback. Drops the now-unused
smart_collections / tag_aliases imports from server.py.

Behaviour-preserving: paths/methods/scopes unchanged (SPA + CLI unaffected).
server.py 3099 → 2967. Boundary test pins leaf-ness, exact route ownership,
server-no-longer-defines, and mounted+auth-guarded (401-not-404).
